Port over make dialog
Make dialog is for the most part legacy functionality, and will
not grow to be contextual. It will keep using the legacy code.
